opt: Separate Insert/Update specific methods from mutationBuilder
Move "ForInsert" and "ForUpdate" methods from mutationBuilder to the
insert.go and update.go files. Leave shared utility methods in the
mutation_builder.go file. This sets the pattern for locating Upsert
and Delete specific files as well.

Release note: None
