+++ pkg/sql/opt/optbuilder/insert.go
@@ -15,9 +15,16 @@
 package optbuilder
 
 import (
+	"sort"
+
 	"github.com/cockroachdb/cockroach/pkg/sql/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt/cat"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt/memo"
+	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/privilege"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/types"
+	"github.com/pkg/errors"
 )
 
 // buildInsert builds a memo group for an InsertOp expression. An input
@@ -149,4 +156,253 @@ func (b *Builder) buildInsert(ins *tree.Insert, inScope *scope) (outScope *scope
 	}
 
 	return mb.outScope
+}
+
+// addTargetNamedColsForInsert adds a list of user-specified column names to the
+// list of table columns that are the target of the Insert operation.
+func (mb *mutationBuilder) addTargetNamedColsForInsert(names tree.NameList) {
+	if len(mb.targetColList) != 0 {
+		panic("addTargetNamedColsForInsert cannot be called more than once")
+	}
+
+	// Add target table columns by the names specified in the Insert statement.
+	mb.addTargetColsByName(names)
+
+	// Ensure that primary key columns are in the target column list, or that
+	// they have default values.
+	mb.checkPrimaryKeyForInsert()
+
+	// Ensure that foreign keys columns are in the target column list, or that
+	// they have default values.
+	mb.checkForeignKeysForInsert()
+}
+
+// checkPrimaryKeyForInsert ensures that the columns of the primary key are
+// either assigned values by the INSERT statement, or else have default/computed
+// values. If neither condition is true, checkPrimaryKeyForInsert raises an
+// error.
+func (mb *mutationBuilder) checkPrimaryKeyForInsert() {
+	primary := mb.tab.Index(cat.PrimaryIndex)
+	for i, n := 0, primary.KeyColumnCount(); i < n; i++ {
+		col := primary.Column(i)
+		if col.Column.HasDefault() || col.Column.IsComputed() {
+			// The column has a default or computed value.
+			continue
+		}
+
+		colID := mb.tabID.ColumnID(col.Ordinal)
+		if mb.targetColSet.Contains(int(colID)) {
+			// The column is explicitly specified in the target name list.
+			continue
+		}
+
+		panic(builderError{pgerror.NewErrorf(pgerror.CodeInvalidForeignKeyError,
+			"missing %q primary key column", col.Column.ColName())})
+	}
+}
+
+// checkForeignKeysForInsert ensures that all composite foreign keys that
+// specify the matching method as MATCH FULL have all of their columns assigned
+// values by the INSERT statement, or else have default/computed values.
+// Alternatively, all columns can be unspecified. If neither condition is true,
+// checkForeignKeys raises an error. Here is an example:
+//
+//   CREATE TABLE orders (
+//     id INT,
+//     cust_id INT,
+//     state STRING,
+//     FOREIGN KEY (cust_id, state) REFERENCES customers (id, state) MATCH FULL
+//   )
+//
+//   INSERT INTO orders (cust_id) VALUES (1)
+//
+// This INSERT statement would trigger a static error, because only cust_id is
+// specified in the INSERT statement. Either the state column must be specified
+// as well, or else neither column can be specified.
+func (mb *mutationBuilder) checkForeignKeysForInsert() {
+	for i, n := 0, mb.tab.IndexCount(); i < n; i++ {
+		idx := mb.tab.Index(i)
+		fkey, ok := idx.ForeignKey()
+		if !ok {
+			continue
+		}
+
+		// This check should only be performed on composite foreign keys that use
+		// the MATCH FULL method.
+		if fkey.Match != tree.MatchFull {
+			continue
+		}
+
+		var missingCols []string
+		allMissing := true
+		for j := 0; j < int(fkey.PrefixLen); j++ {
+			indexCol := idx.Column(j)
+			if indexCol.Column.HasDefault() || indexCol.Column.IsComputed() {
+				// The column has a default value.
+				allMissing = false
+				continue
+			}
+
+			colID := mb.tabID.ColumnID(indexCol.Ordinal)
+			if mb.targetColSet.Contains(int(colID)) {
+				// The column is explicitly specified in the target name list.
+				allMissing = false
+				continue
+			}
+
+			missingCols = append(missingCols, string(indexCol.Column.ColName()))
+		}
+		if allMissing {
+			continue
+		}
+
+		switch len(missingCols) {
+		case 0:
+			// Do nothing.
+		case 1:
+			panic(builderError{errors.Errorf(
+				"missing value for column %q in multi-part foreign key", missingCols[0])})
+		default:
+			sort.Strings(missingCols)
+			panic(builderError{errors.Errorf(
+				"missing values for columns %q in multi-part foreign key", missingCols)})
+		}
+	}
+}
+
+// addTargetTableColsForInsert adds up to maxCols columns to the list of columns
+// that will be set by an INSERT operation. Columns are added from the target
+// table in the same order they appear in its schema. This method is used when
+// the target columns are not explicitly specified in the INSERT statement:
+//
+//   INSERT INTO t VALUES (1, 2, 3)
+//
+// In this example, the first three columns of table t would be added as target
+// columns.
+func (mb *mutationBuilder) addTargetTableColsForInsert(maxCols int) {
+	if len(mb.targetColList) != 0 {
+		panic("addTargetTableColsForInsert cannot be called more than once")
+	}
+
+	numCols := 0
+	for i, n := 0, mb.tab.ColumnCount(); i < n && numCols < maxCols; i++ {
+		// Skip hidden columns.
+		if mb.tab.Column(i).IsHidden() {
+			continue
+		}
+
+		mb.addTargetCol(i)
+		numCols++
+	}
+
+	// Ensure that the number of input columns does not exceed the number of
+	// target columns.
+	mb.checkNumCols(len(mb.targetColList), maxCols)
+}
+
+// buildInputForInsert constructs the memo group for the input expression and
+// constructs a new output scope containing that expression's output columns.
+func (mb *mutationBuilder) buildInputForInsert(inScope *scope, inputRows *tree.Select) {
+	// If there are already required target columns, then those will provide
+	// desired input types. Otherwise, input columns are mapped to the table's
+	// non-hidden columns by corresponding ordinal position. Exclude hidden
+	// columns to prevent this statement from writing hidden columns:
+	//
+	//   INSERT INTO <table> VALUES (...)
+	//
+	// However, hidden columns can be written if the target columns were
+	// explicitly specified:
+	//
+	//   INSERT INTO <table> (...) VALUES (...)
+	//
+	var desiredTypes []types.T
+	if len(mb.targetColList) != 0 {
+		desiredTypes = make([]types.T, len(mb.targetColList))
+		for i, colID := range mb.targetColList {
+			desiredTypes[i] = mb.md.ColumnMeta(colID).Type
+		}
+	} else {
+		desiredTypes = make([]types.T, 0, mb.tab.ColumnCount())
+		for i, n := 0, mb.tab.ColumnCount(); i < n; i++ {
+			tabCol := mb.tab.Column(i)
+			if !tabCol.IsHidden() {
+				desiredTypes = append(desiredTypes, tabCol.DatumType())
+			}
+		}
+	}
+
+	mb.outScope = mb.b.buildSelect(inputRows, desiredTypes, inScope)
+
+	if len(mb.targetColList) != 0 {
+		// Target columns already exist, so ensure that the number of input
+		// columns exactly matches the number of target columns.
+		mb.checkNumCols(len(mb.targetColList), len(mb.outScope.cols))
+	} else {
+		// No target columns have been added by previous steps, so add columns
+		// that are implicitly targeted by the input expression.
+		mb.addTargetTableColsForInsert(len(mb.outScope.cols))
+	}
+
+	// Loop over input columns and:
+	//   1. Type check each column
+	//   2. Assign name to each column
+	//   3. Add id of each column to the insertColList
+	for i := range mb.outScope.cols {
+		inCol := &mb.outScope.cols[i]
+		ord := mb.tabID.ColumnOrdinal(mb.targetColList[i])
+
+		// Type check the input column against the corresponding table column.
+		checkDatumTypeFitsColumnType(mb.tab.Column(ord), inCol.typ)
+
+		// Assign name of input column. Computed columns can refer to this column
+		// by its name.
+		inCol.name = tree.Name(mb.md.ColumnMeta(mb.targetColList[i]).Alias)
+
+		// Map the ordinal position of each table column to the id of the input
+		// column which will be inserted into that position.
+		mb.insertColList[ord] = inCol.id
+	}
+}
+
+// buildEmptyInput constructs a new output scope containing a single row VALUES
+// expression with zero columns.
+func (mb *mutationBuilder) buildEmptyInput(inScope *scope) {
+	mb.outScope = inScope.push()
+	mb.outScope.expr = mb.b.factory.ConstructValues(memo.ScalarListWithEmptyTuple, opt.ColList{})
+}
+
+// addDefaultAndComputedColsForInsert wraps an Insert input expression with
+// Project operator(s) containing any default (or nullable) and computed columns
+// that are not yet part of the target column list. This includes mutation
+// columns, since they must always have default or computed values.
+//
+// After this call, the input expression will provide values for every one of
+// the target table columns, whether it was explicitly specified or implicitly
+// added.
+func (mb *mutationBuilder) addDefaultAndComputedColsForInsert() {
+	// Add any missing default and nullable columns.
+	mb.addSynthesizedCols(
+		mb.insertColList,
+		func(tabCol cat.Column) bool { return !tabCol.IsComputed() },
+	)
+
+	// Add any missing computed columns. This must be done after adding default
+	// columns above, because computed columns can depend on default columns.
+	mb.addSynthesizedCols(
+		mb.insertColList,
+		func(tabCol cat.Column) bool { return tabCol.IsComputed() },
+	)
+}
+
+// buildInsert constructs an Insert operator, possibly wrapped by a Project
+// operator that corresponds to the given RETURNING clause.
+func (mb *mutationBuilder) buildInsert(returning tree.ReturningExprs) {
+	private := memo.MutationPrivate{
+		Table:       mb.tabID,
+		InsertCols:  mb.insertColList,
+		NeedResults: returning != nil,
+	}
+	mb.outScope.expr = mb.b.factory.ConstructInsert(mb.outScope.expr, &private)
+
+	mb.buildReturning(returning)
 }
